feat: rename telemetrysrc package to reader
