+++ internal/reader/reader.go
@@ -1,4 +1,4 @@
-package telemetrysrc
+package reader
 
 type Reader interface {
 	Read() (int, []byte, error)
